pyproject.toml
[build-system]
requires      = ["setuptools~=62.6", "wheel~=0.37.1"]
build-backend = "setuptools.build_meta"


[project]
name        = "datoso"
description = "Python command line tool to download and organize your Rom Dat files."
readme      = "README.md"
requires-python = ">=3.10"
license     = {text = "MIT License"}
authors     = [
    {name = 'Lacides Miranda', email = 'laromicas@hotmail.com'},
]
keywords = ["emulators", "roms"]
classifiers = [
    'Development Status :: 3 - Alpha',
    "Environment :: Console",
    'License :: OSI Approved :: MIT License',
    "Operating System :: POSIX :: Linux",
    'Programming Language :: Python :: 3',
    'Topic :: System :: Emulators',
]
dependencies = [
    "tinydb>=4.7.0",
    "pydantic>=1.9.2",
    "python-dateutil>=2.8.0",
    "xmltodict>=0.13.0",
    "internetarchive>=3.0.0",
    "tabulate>=0.9.0",
    "orjson>=3.8.10",
]
dynamic = ["version"]


[project.optional-dependencies]
selenium = [
    "selenium==4.3.0",
    ]

nointro = [
    "datoso_nointro>=0.1.0",
    ]

redump = [
    "datoso_redump>=0.1.0",
    ]

dev = [
    "ruff>=0.0.261",
    ]

[project.urls]
"Source Code"       = "https://github.com/laromicas/datoso"

[project.scripts]
datoso = "datoso.__main__:main"

[tool.setuptools]
packages = ["datoso"]
license-files = ["LICENSE"]
include-package-data = true

[tool.setuptools.dynamic]
version = {attr = "datoso.__version__"}

[tool.ruff]
extend-select = ["C4", "SIM", "TCH"]
ignore = []
fixable = ["F541"]
# fixable = ["A", "ANN", "ARG", "B", "BLE", "C", "COM", "D", "DTZ", "E", "EM", "ERA", "EXE", "F", "FBT", "G", "I", "ICN", "INP", "ISC", "N", "PD", "PGH", "PIE", "PL", "PT", "PTH", "Q", "RET", "RUF", "S", "SIM", "T", "TCH", "TID", "TRY", "UP", "W", "YTT"]
line-length = 120
show-fixes = true
target-version = "py310"
ignore-init-module-imports = true

# TODO Maybe change license to "BSD 3-Clause License"
